// src/types/Computer.tsx
export type Computer = {
  case: {
    name: string;
    price: number;
  };
  motherboard: {
    name: string;
    price: number;
  };
  cpu: {
    name: string;
    price: number;
  };
  gpu: {
    name: string;
    price: number;
  };
  ram: {
    name: string;
    price: number;
  };
  storage: {
    name: string;
    price: number;
  };
  cooling: {
    name: string;
    price: number;
  };
  psu: {
    name: string;
    price: number;
  };
};